+++ sb_pipe/utils/python/copasi_parser.py
